ITO-000 create enpoint down file.

+++ src/main/java/onlinejudge/resource/controller/ResourceController.java
@@ -5,7 +5,9 @@ import java.io.IOException;
 import java.math.BigInteger;
 import java.net.Authenticator.RequestorType;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Value;
@@ -21,11 +23,22 @@ import onlinejudge.dto.file.MyResource;
 
 @Controller
 public class ResourceController {
-	@Value("${path.problem}")
-	String problemPath;
 	
-	@Value("${path.submit}")
-	String submitPath;
+	
+	Map<String, String> paths;
+	
+	
+	public ResourceController(@Value("${path.problem}") String problemPath,@Value("${path.submit}") String submitPath) {
+		super();
+		paths = new HashMap<>();
+		paths.put(MyResource.RESOURCE_TYPE_PROBLEM, problemPath);
+		paths.put(MyResource.RESOURCE_TYPE_TESTCASE_INPUT, problemPath);
+		paths.put(MyResource.RESOURCE_TYPE_TESTCASE_OUTPUT, problemPath);
+		paths.put(MyResource.RESOURCE_TYPE_SUBMIT, submitPath);
+	}
+
+	private void addPathToMap(){
+	}
 	
 //	@RequestMapping({"/","/about"})
 //	public String about(){
@@ -39,43 +52,45 @@ public class ResourceController {
 		group.add(resource);
 		return group;
 	}
-	
-	@RequestMapping(value = "/upfile", method=RequestMethod.POST)
+	/**
+	 * #resource-001
+	 * @param groupResource
+	 * @return
+	 */
+	@RequestMapping(value = "/upfiles", method=RequestMethod.POST)
 	public @ResponseBody MyResponse upfile(@RequestBody GroupResource groupResource){
 		MyResponse myResponse = null;
 		List<String> listBackupFileName = new ArrayList<>();
 		List<String> listNewFileName = new ArrayList<>();
 		try{
 			for(MyResource resource: groupResource.getListResource()){
-				
-				if(MyResource.RESOURCE_TYPE_PROBLEM.equals(resource.getResourceType())){
-					
-					File fileOutput = new File(problemPath + resource.getFileName()); // File will be wrote data into
-					boolean isNewFile = !fileOutput.exists();
-					if(!isNewFile){ // need backup
-						File backupFile = new File(fileOutput.getAbsolutePath()+"." + System.currentTimeMillis());
-						try{
-							FileUtils.copyFile(fileOutput, backupFile);
-							listBackupFileName.add(backupFile.getAbsolutePath());
-						}catch (IOException e){
-							e.printStackTrace();
-							throw new ExceptionNeedRollback(e.getMessage());
-						}
-					}
-					
-					try {
-						FileUtils.copyInputStreamToFile(resource.inputStream(),fileOutput);
-						if(isNewFile)
-							listNewFileName.add(fileOutput.getAbsolutePath());
-					} catch (IOException e) {
+				File fileOutput = new File(getPath(resource)); // File will be wrote data into
+				boolean isNewFile = !fileOutput.exists();
+				if(!isNewFile){ // need backup
+					File backupFile = new File(fileOutput.getAbsolutePath()+"." + System.currentTimeMillis());
+					try{
+						FileUtils.copyFile(fileOutput, backupFile);
+						listBackupFileName.add(backupFile.getAbsolutePath());
+					}catch (IOException e){
 						e.printStackTrace();
 						throw new ExceptionNeedRollback(e.getMessage());
 					}
 				}
+				
+				try {
+					FileUtils.copyInputStreamToFile(resource.inputStream(),fileOutput);
+					if(isNewFile)
+						listNewFileName.add(fileOutput.getAbsolutePath());
+				} catch (IOException e) {
+					e.printStackTrace();
+					throw new ExceptionNeedRollback(e.getMessage());
+				}
 			}
 			
-			for (String path : listBackupFileName) {
-				FileUtils.deleteQuietly(new File(path));
+			if(!groupResource.isKeepBackup()){
+				for (String path : listBackupFileName) {
+					FileUtils.deleteQuietly(new File(path));
+				}
 			}
 			myResponse = MyResponse.builder().success().build();
 		}catch(ExceptionNeedRollback e){
@@ -95,7 +110,36 @@ public class ResourceController {
 		}
 		return myResponse;
 	}
+	/**
+	 * #resource-002
+	 * @param groupResource
+	 * @return
+	 */
+	@RequestMapping("/downfiles")
+	public @ResponseBody GroupResource download( @RequestBody GroupResource groupResource){
+		File file = null;
+		byte[] data = null;
+		for (MyResource resource : groupResource.getListResource()) {
+			 file = FileUtils.getFile(getPath(resource));
+			if(file.exists()){
+				try {
+					data = FileUtils.readFileToByteArray(file);
+					resource.setData(data);
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			}
+		}
+		return groupResource;
+	}
+	
+	private String getPath(MyResource resource){
+		return paths.get(resource.getResourceType()) + resource.getFileName();
+	}
 }
+
+
+
 class ExceptionNeedRollback extends Exception{
 	private static final long serialVersionUID = -1694883875581285891L;
